backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Async & Background Tasks
celery==5.4.0
redis==5.2.0

# AI & Media Processing
groq==0.11.0
ffmpeg-python==0.2.0
google-generativeai>=0.8.0

# Cloud Storage
boto3==1.35.0  # Cloudflare R2 (S3-compatible)
supabase==2.9.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator>=2.0.0
PyJWT>=2.8.0

# Development
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
